rename: components 폴더 내 파일 pages 폴더로 이동
components 폴더 내 라우팅 적용할 페이지 컴포넌트 파일 폴더 이동
